+++ packages/fighting-design/textarea/src/props.ts
@@ -55,9 +55,9 @@ export const Props = {
    *
    * @see resize https://developer.mozilla.org/zh-CN/docs/Web/CSS/resize
    * @values none both horizontal vertical
-   * @default none
+   * @default null
    */
-  resize: setStringProp<TextareaResize>('none', (val: TextareaResize): boolean => {
+  resize: setStringProp<TextareaResize>(null, (val: TextareaResize): boolean => {
     return (['none', 'both', 'horizontal', 'vertical'] as const).includes(val)
   }),
   /** 是否可清空 */